requirements.txt
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
exceptiongroup==1.1.3
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
iniconfig==2.0.0
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
outcome==1.3.0.post0
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.4.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.23.1
trio-websocket==0.11.1
uritemplate==4.1.1
urllib3==2.1.0
wsproto==1.2.0
